+++ CHANGELOG.md
@@ -71,6 +71,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Report meaningful error message instead of generic HTTP/500 internal error
   when authenticating with JWT token on unexisting remote HTTP instance (#137).
 - Fix crash with when watching streamed tasks output with HTTP REST API (#138).
+- Avoid buffering with HTTP response headers on reverse proxies (#139).
 
 ## [2.0.0] - 2023-05-05